Make profanity keybindings like vim and qutebrowser
This is a strange mix of vim-like and qutebrowser (which is also
vim-like) keybindings.  Since qutebrowser and profanity both share a
tab-centric workflow, the keybinding to switch tabs is easy to use ("J"
and "K"), while C-{f,b} is retained for similarity with vi(m), less, and
other unix tools.  The keybinding for paging the roster/occupants list
is not going to be used much, so g<CHAR> is warranted.