sxhkd: Add keybind to swap arbitrary nodes
I already have keybinds for swapping with bigger nodes and with adjacent
nodes (which are pretty common operations), but swapping with arbitrary
nodes is also useful.

The marked node is not unmarked for now.  I'd like to see how I use this
keybind and whether I do anything with the marked node.