emacs: Only use buffer for the buffer/file embark target
Some of the commands work on buffers but are passed a file as the
argument, which were absolute paths for me.  Buffer names aren't
generally paths (let alone absolute ones), so they didn't work.

Using just the buffer works well enough for commands which need files
or those which need buffers.  At least when there are no name
conflicts.